irqchip, gicv3-its, cma: Use CMA for allocation of large device tables
The gicv3-its device table may have a size of up to 16MB. With 4k
pagesize the maximum size of memory allocation is 4MB. Use CMA for
allocation of large tables.

We use the device managed version of dma_alloc_coherent(). Thus, we
don't need to release it manually on device removal.

Signed-off-by: Robert Richter <rrichter@cavium.com>
Message-Id: <20170201173833.12630-10-rrichter@cavium.com>
Signed-off-by: Robert Richter <rrichter@cavium.com>