arm64: Disable preemption before entering panic()
There is an issue with CONFIG_PREEMPT=y and "isolcpus" or
CONFIG_NO_HZ_FULL_ALL, where after a kernel panic the user space
programs are still running. This makes it difficult for the hardware
watchdog to trigger as the user space daemon can update the timer
through ioctls. By disabling the preemption before entering the panic()
the issue is not seen. It is a temporary workaround until actual issue
is fixed

Signed-off-by: Radha Mohan Chintakuntla <rchintakuntla@cavium.com>