octeontx: update domain creation function
Update domain creation function in order to allow creation of a domain
without BGX or LBK ports. Domain creation function also validates if
there is at least 1 SSO, SSOW and PKI unit in the domain. It also
validates that the number of PKO units is equal to number of selected
BGX and LBK ports. If not, command returns an error.

Signed-off-by: Lukasz Bartosik <lb@semihalf.com>
Signed-off-by: Stanislaw Kardach <kda@semihalf.com>
Signed-off-by: Yury Norov <ynorov@caviumnetworks.com>