PCI: Move pci-octeontx-ecam driver before the pci-host-generic
The pci-octeontx-ecam driver emulates the PKI VF space and so uses a
custom compatible string but we want the ability to use upstream kernels
too on CN83xx which doesn't have pci-octeontx-ecam driver. The upstream
kernels use the pci-host-generic driver. We pass both the compatible
strings in the device tree so that same firmware can be used for
multiple kernels. Moving the pci-octeontx-ecam driver up in the Makefile
makes it to be called before the pci-host-generic driver, thereby giving
the advantage.

Signed-off-by: Radha Mohan Chintakuntla <rchintakuntla@cavium.com>
Signed-off-by: Yury Norov <ynorov@caviumnetworks.com>