octeontx-bgx: Reduce BGX TX FIFO threashold for 10GE interfaces
The threashold value size should not be bigger then either BGX TX FIFO
or PKO TX FIFO sizes. For XFI mode they are 12KB and 2.5KB accordingly.
The previous threashold value 0x100 was equvalent to 4KB, which was
causing PKO to freeze, when passing big packets. New value is setup to
0x80 (2KB).

Signed-off-by: Yuri Tolstov <yuri.tolstov@cavium.com>