Add try_get_task_struct_on_cpu() to scheduler for task isolation
Task isolation wants to be able to verify that a remote core is
running an isolated task to determine if it should generate a
diagnostic, and also possibly interrupt it.

This API returns a pointer to the task_struct of the task that
was running on the specified core at the moment of the request;
it uses try_get_task_struct() to increment the ref count on the
returned task_struct so that the caller can examine it even if
the actual remote task has already exited by that point.

Signed-off-by: Chris Metcalf <cmetcalf@mellanox.com>
Signed-off-by: Yury Norov <ynorov@caviumnetworks.com>