net: thunderx: Enable independent Rx packet size limit setting per LMAC
As of now MTU is enforced only on the transmit side, but on Rx
path packets of size 'NIC_HW_MAX_FRS' are allowed. This patch
changes the behavior to restrict Rx packets also to MTU size.
For this a separate PKIND is assigned to each LMAC and max length
in PKIND is set to MTU.

Signed-off-by: Sunil Goutham <sgoutham@cavium.com>
Signed-off-by: Maciej Czekaj <maciej.czekaj@cavium.com>