crypto: cpt - move domain code to separate file
We want to keep all code which will not be upstreamed in separate files
because CPT driver version for upstreaming is kept in cpt-driver repo.
In order to be able to port changes between CPT drivers in cpt-driver
and kernel repos we want them to be as much common as possible.

Signed-off-by: Lukasz Bartosik <lukasz.bartosik@cavium.com>