crypto: cpt - add support for firmware in a blob
When CPT driver is build into a kernel we need a way to provide it with
firmware because at the moment of CPT initialization rootfs is not
ready. For that purpose we will use a blob. When CPT driver is compiled
as a module we will use request_firmware from kernel API to load
firmware with a fallback to blob in case request_firmware fails.

Signed-off-by: Lukasz Bartosik <lukasz.bartosik@cavium.com>