arm64: flush pipeline early at exit of extended quiescent state
This series enables delaying of kernel memory synchronization
for CPUs running in extended quiescent state till the exit of
that state.

For memory synchronization on all CPUs we use IPI mechanism.
If IPI is fired on CPU, it calls isb() during gic_handle_irq(),
and text patch framework implicitly relies on that isb() (see
aarch64_insn_patch_text()).

So if we don't fire IPI for EQS CPUs anymore, we should add ISB
early in EQS exit path.

There are 2 such paths. One starts in do_idle() loop, and other
in el0_svc entry. For do_idle(), isb() is added in
arch_cpu_idle_exit() hook which is not used in arm64 at now.
And for SVC handler, isb is called in el0_svc_naked.

Suggested-by: Will Deacon <will.deacon@arm.com>