(SDK only) spidev: expand device list, add generic "linux,spidev"
This should go in Cavium SDK, but not to kernel.org.

Changing the kernel to add compatible devices seems wrong,
but that's what the spidev code implies - logging a warning
if compatible="spidev" is listed in device-tree.

Two SPI-driven telephony devices (SLIC and framer) are added to
compatible-id list in the spirit of kernel.org's 956b200a.

In practice many distributions have found this to be unworkable,
and added a "linux,spidev" generic id, so that's added too,
to avoid having to rework generic thunder kernel when a new SPI
device is added, which can be managed from usermode by spidev.

One of these changes should go upstream, but perhaps not both,
so that devtree-described SPI targets can be accessed from
usermode via kernel spidev driver

Signed-off-by: Peter Swain <pswain@cavium.com>