soc: Move SOC entry to under IOMMU
This is a temporary change (not for upstreaming) until
below issue is debugged and fixed properly.

Due to changes in IOMMU infrastructure in 4.14 kernel,
SMMUv3 driver is being probed after PCI devices enumeration
is done. Earlier on 4.9 kernel it used to happen before
ECAM enumeration.

And since in drivers/makefile IOMMU comes after SOC, OcteonTX2
RVU AF driver is being probed before SMMUv3 is initialized
resulting in lots of SMMU errors while kernel is booting.
Though nothing is broken, it looks pretty ackward and raising
false alarms. So moved SOC entry in Makefile to under IOMMU,
so that SMMUv3 driver is initialized first.

Signed-off-by: Sunil Goutham <sgoutham@marvell.com>
Signed-off-by: Radha Mohan Chintakuntla <radhac@marvell.com>