mmc: cavium: forbid unaligned DMA
The DMA engine used in Cavium SoCs multi-block MMC transfers
requires 64bit alignment. Force failure on unaligned requests.

It should be possible for unaligned requests to fall back to
iterating over the direct-access of single-block access,
but that seems to provoke sg_miter confusion when crossing
block boundaries which are not page boundaries.

Returning -EINVAL for now, consider that optimization later.

This gives a clean UNSUPPORTED_BY_HOST to mmc_test 13/14.

Also moved the error = -EINVAL earlier, so it must be actively
overwritten by successful completion

Signed-off-by: Peter Swain <peter.swain@marvell.com>