mm: fix COW faults after mlock()
After mlock() on newly mmap()ed area page faults observed.

The problem is that populate_vma_page_range() doesn't set FOLL_WRITE
flag for shared memory arguing that like:
/*
 * We want to touch writable mappings with a write fault in order
 * to break COW, except for shared mappings because these don't COW
 * and we would not want to dirty them for nothing.
 */

But they are actually COWed, and the easiest way to avoid it is to
set VM_WRITE flag for shared mappings as well as for private ones.

Signed-off-by: Yury Norov <ynorov@caviumnetworks.com>