lib/test_bitmap: delete meaningless test for bitmap_cut
One of bitmap_cut tests passes it with:
	nbits = BITS_PER_LONG;
	first = BITS_PER_LONG;
	cut = BITS_PER_LONG;

This test is useless because the range to cut is not inside the
bitmap. This should normally raise an error, but bitmap_cut() is
void and returns nothing.

The test just checks that the memory is not touched by bitmap_cut(),
which is not correct, because if a function is passed with wrong
parameters, it's too optimistic to expect a correct, or even sane
behavior.

Now that we have bitmap_check_params(), there's a tool to avoid such
things in real code, and we can just drop the test.