cpumask: switch for_each_cpu{,_not} to use for_each_bit()
The difference between for_each_cpu() and for_each_set_bit()
is that the latter uses cpumask_next() instead of find_next_bit(),
and so calls cpumask_check().

This check is useless because the iterator is not provided by user.
It generates many false-positives for the very last iteration of
for_each_cpu(), if the cpumask_check() is fixed, like in the following
patches.

Signed-off-by: Yury Norov <yury.norov@gmail.com>