bitmap: improve small_const case for find_next() functions
find_next_bit() and friends use small_const_nbits() to detect possibility
of compile-time optimization. It works well for nbits up to BITS_PER_LONG,
but doesn't work when offset and nbits belong to 2nd, 3rd or any other word
except the 1st one.

When both offset and size are within the same word, we can make a
small_const optimization because the whole business is still to fetch a
single word, and mask unneeded head and tail bits. This all can be done at
compile time.

New small_const_nbits_off() macro is doing that work, and is used to
replace small_const_nbits() in find_next_bit() functions.

Signed-off-by: Yury Norov <yury.norov@gmail.com>