UPSTREAM PENDING: ASoC: tegra: Fix SFC conversion for few rates
Sample rate conversions for rates greater than 48kHz are
found to be failing.
x->y resampling fails where either x or y is greater than 48kHz.

This happens because, tegra210_sfc_rate_to_idx() returns incorrect
index for rates greater than 48kHz. This actually depends on the
tegra210_sfc_rates[] array and it is not in sync with TX/RX register
spec.
To be precise, 64kHz entry is missing in above array which is otherwise
present in the spec. Due to this wrong index is returned and this
results in incorrect programming of coefficients.

To fix this,
align the tegra210_sfc_rates[] array with register
spec and thus add 64kHz entry to it.
also, the coefficient table is updated to reflect that none of the
conversions are supported for 64kHz.

Bug 3896803

Change-Id: I46def4f17a0818f009453a0d9bdbfbd6715e5100
Signed-off-by: Sheetal <sheetal@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-stable/+/2856333
(cherry picked from commit 7cd2f54257df1316788bd311081c69feaf5e909a)
Reviewed-on: https://git-master.nvidia.com/r/c/linux-stable/+/2886661
Reviewed-by: Mohan Kumar D <mkumard@nvidia.com>
Reviewed-by: Sharad Gupta <sharadg@nvidia.com>
GVS: Gerrit_Virtual_Submit <buildbot_gerritrpt@nvidia.com>
Signed-off-by: Shanker Donthineni <sdonthineni@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-stable/+/2890184
Reviewed-by: Rohit Khanna <rokhanna@nvidia.com>