sched/topology: update previous mask in for_each_numa_hop_mask() automatically
There are 2 users of for_each_numa_hop_mask(). One is in-tree Mellanox's
driver (although, recently converted), and another one is in-review
Intel's ICE driver:

https://lore.kernel.org/lkml/20230216145455.661709-1-pawel.chmielewski@intel.com/

which I hope will soon be merged.

They both track a mask associated with previous hop. It looks like it might
be a typical use-case, so it worth to handle a mask for previous hop inside
for_each_numa_hop_mask().

Signed-off-by: Yury Norov <yury.norov@gmail.com>