sched/topology: add for_each_numa_cpu() macro
for_each_cpu() is widely used in the kernel, and seemingly, it would be
beneficial to create a NUMA-aware version of the macro.

Recently added for_each_numa_hop_mask() does the job, but switching
existing codebase to use that macro is not a straightforward process.
The for_each_numa_cpu() is designed to be as similar to the for_each_cpu()
as possible. To make the code NUMA-aware, one must just pass a couple
iterator variables inside new macro, and it takes care of the rest.

The tradeof of for_each_numa_cpu() is that it is implemented as 2 nested
for-loops, and therefore it's unsafe to use 'break' in the body of the
loop. But 'goto' works just well, and it's worth to have a simple
NUMA-aware replacement for the for_each_cpu().

Signed-off-by: Yury Norov <yury.norov@gmail.com>