sched/topology: add for_each_numa_cpu() macro
for_each_cpu() is widely used in the kernel, and it would be beneficial
to create a NUMA-aware version of the macro.

Recently added for_each_numa_hop_mask() does the job, but switching
existing codebase to use that macro is not a straightforward process.

New for_each_numa_cpu() is designed to be similar to the for_each_cpu().
It allows to convert existing code to NUMA-aware as simple as adding a
hop iterator variable and passing it inside new macro. for_each_numa_cpu()
takes care of the rest.

The trade-off for this approach is that the new macro is implemented as
a nested for-loop, and therefore it's unsafe to use 'break' in the body
of the loop.

But 'goto' works just well, and it's worth to have a simple NUMA-aware
replacement for the for_each_cpu().

At the moment, we have 2 users of NUMA-aware enumerators. One is
Melanox's in-tree driver, and one Intel's in-review driver:

https://lore.kernel.org/lkml/20230216145455.661709-1-pawel.chmielewski@intel.com/

Both real-life examples follow the same pattern:

	for_each_numa_hop_mask(cpus, prev, node) {
 		for_each_cpu_andnot(cpu, cpus, prev) {
 			if (cnt++ == max_num)
 				goto out;
 			do_something(cpu);
 		}
		prev = cpus;
 	}

With the new macro, it would look like this:

	for_each_numa_cpu(cpu, hop, node, cpu_possible_mask) {
		if (cnt++ == max_num)
			goto out;
		do_something(cpu);
 	}

Switching existing codebase to NUMA-aware version now shouldn't be a
brainteaser. Considering the example above, NUMA-ignorant code may
look like:

	cpumask_t cpus = get_mask();
	int cnt = 0, cpu;

	for_each_cpu(cpu, cpus) {
		if (cnt++ == max_num)
			goto out;
		do_something(cpu);
 	}

Converting it to NUMA-aware version would be as simple as:

	cpumask_t cpus = get_mask();
	int node = get_node();
	int cnt = 0, cpu;
	int hop = 0;

	for_each_numa_cpu(cpu, hop, node, cpus) {
		if (cnt++ == max_num)
			goto out;
		do_something(cpu);
 	}

Another advantage of for_each_numa_cpu() over for_each_numa_hop_mask()
is that it introduces a 'hop' parameter with the clear meaning of NUMA
distance, and doesn't make people not familiar to enumerator internals
bothering with current and previous masks machinery.

Signed-off-by: Yury Norov <yury.norov@gmail.com>