sched/topology: introduce sched_numa_find_next_cpu()
Similarly to sched_numa_find_nth_cpu(), add sched_numa_find_next_cpu().
The function searches for the next CPU in a given cpumask according to
NUMA topology, so that if a given CPU is the last cpu in a given hop,
the next CPU will be the first cpu in a next hop.