numa: generalize numa_map_to_online_node()
The function is in fact searching for the nearest node to a given one,
based on a N_ONLINE state. This is a common pattern to search for a
nearest node. Particularly, the following patch adds such traversing
in a scheduler code.

This patch converts numa_map_to_online_node(node) to
numa_nearest_node(node, state), so that others won't need opencode the
logic.