bitmap: separate handling of identity and remapping parts in bitmap_remap()
For unset bits in 'old' map (identity mapping), 'src' bits must get copied
to 'dst'. Doing that separately from remapping in a for-loop has some
advantages:
 - implicitly initialize 'dst' without calling bitmap_zero();
 - optimize performance of handling identity parts, because per-word
   bitmap_andnot() is faster than per-bit set_bit() in a for-loop;
 - make inner part of the loop unconditional;
 - when 'old' map is empty, new logic simiply skips the for-loop part.

While here, replase set_bit() with a non-atomic version, because the
function is non-atomic anyways. If atomicity required, it should be
handled by user.

Signed-off-by: Yury Norov <yury.norov@gmail.com>