bitmap: don't count bits from the beginning in bitmap_remap()
In the remapping part, we count bits in 'old' and 'new' masks from the
beginning for every bit. Complexity of this is O(N^2). We can do it
better in O(N), if switch bitmap_remap() to using _from versions of
bitmap_weight() and find_nth_bit().

On kvm/x86_64 it works 9x faster than the current version for the
1000-bit bitmap in lib/test_bitmap.

Signed-off-by: Yury Norov <yury.norov@gmail.com>