lib/find: add find_and_clear_bit()
Add a helper around test_and_clear_bit() that allows to search for a set
bit and clear atomically.

The pattern to be replaced is:

	do {
		bit = find_first_bit(bitmap, size);
		if (bit >= size)
			return size;
	} while (!test_and_clear_bit(bit, bitmap));

This not only a handy helper, it also resolves a non-trivial issue of
using non-volatile find_bit() together with atomic test_and_clear_bit().

The problem is that find_bit() implies that the bitmap is a regular
non-volatile piece of memory, and compiler is allowed to use all
optimization techniques like re-fetching memory instead of caching it.

find_first_bit() is implemented like this:
      for (idx = 0; idx * BITS_PER_LONG < sz; idx++) {
              val = addr[idx];
              if (val) {
                      sz = min(idx * BITS_PER_LONG + __ffs(val), sz);
                      break;
              }
      }

On register-memory architectures, like x86, compiler may decide to
access memory twice - first time to compare against 0, and second time
to fetch its value to pass it to __ffs().

When running find_first_bit() on volatile memory, the memory may get
changed in-between, and for instace, it may lead to passing 0 to
__ffs(), which is undefined, and potentionally dangerous call.

find_and_clear_bit(), as a wrapper around test_and_clear_bit(),
naturally treats underlying bitmap as a volatile memory, and prevents
comiler from such optimizations.

Now that we have KCSAN, which is designed to catch exactly this
situations, it warns against undercover memory modifications. We
can use this tool to reveal improper usage of find_first_bit(),
and convert it to find_and_clear_bit() as appropriate.

Signed-off-by: Yury Norov <yury.norov@gmail.com>